requirements.txt
#pip install Flask Flask-CORS pymongo Flask-MongoEngine python-dotenv

Flask
Flask-CORS
pymongo
Flask-MongoEngine
python-dotenv